worker: DD move BitStream to Utils
