+++ worker/meson.build
@@ -68,6 +68,7 @@ common_sources = [
   'src/Settings.cpp',
   'src/Worker.cpp',
   'src/ChannelMessageRegistrator.cpp',
+  'src/Utils/BitStream.cpp',
   'src/Utils/Crypto.cpp',
   'src/Utils/File.cpp',
   'src/Utils/IP.cpp',
@@ -130,7 +131,6 @@ common_sources = [
   'src/RTC/UdpSocket.cpp',
   'src/RTC/WebRtcServer.cpp',
   'src/RTC/WebRtcTransport.cpp',
-  'src/RTC/Codecs/BitStream.cpp',
   'src/RTC/Codecs/H264.cpp',
   'src/RTC/Codecs/H264_SVC.cpp',
   'src/RTC/Codecs/VP8.cpp',
